+++ examples/web-rest-openapi-funcs.ps1
@@ -1,0 +1,13 @@
+$path = Split-Path -Parent -Path (Split-Path -Parent -Path $MyInvocation.MyCommand.Path)
+Import-Module "$($path)/src/Pode.psm1" -Force -ErrorAction Stop
+
+Start-PodeServer {
+    Add-PodeEndpoint -Address localhost -Port 8080 -Protocol Http
+
+    Enable-PodeOpenApi -Title 'OpenAPI Example' -RouteFilter '/api/*' -RestrictRoutes
+    Enable-PodeOpenApiViewer -Type Swagger -DarkMode
+    Enable-PodeOpenApiViewer -Type ReDoc
+
+    #ConvertTo-PodeRoute -Path '/api' -Commands @('Get-ChildItem', 'New-Item')
+    ConvertTo-PodeRoute -Path '/api' -Module Pester
+}