+++ src/Private/SignalServer.ps1
@@ -6,26 +6,17 @@ function Start-PodeSignalServer
     # work out which endpoints to listen on
     $endpoints = @()
     @(Get-PodeEndpoints -Type Ws) | ForEach-Object {
-        # get the protocol
-        $_protocol = (Resolve-PodeValue -Check $_.Ssl -TrueValue 'wss' -FalseValue 'ws')
-
         # get the ip address
         $_ip = [string]($_.Address)
         $_ip = (Get-PodeIPAddressesForHostname -Hostname $_ip -Type All | Select-Object -First 1)
         $_ip = (Get-PodeIPAddress $_ip)
 
-        # get the port
-        $_port = [int]($_.Port)
-        if ($_port -eq 0) {
-            $_port = (Resolve-PodeValue $_.Ssl -TrueValue 9443 -FalseValue 9080)
-        }
-
         # add endpoint to list
         $endpoints += @{
             Address = $_ip
-            Port = $_port
+            Port = $_.Port
             Certificate = $_.Certificate.Raw
-            HostName = "$($_protocol)://$($_.HostName):$($_port)/"
+            HostName = $_.Url
         }
     }